+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dyrected/nuxt
 
+## 2.5.33
+
+### Patch Changes
+
+- @dyrected/admin@2.5.33
+- @dyrected/core@2.5.33
+- @dyrected/sdk@2.5.33
+- @dyrected/vue@2.5.33
+
 ## 2.5.32
 
 ### Patch Changes