+++ packages/storage-cloudinary/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dyrected/storage-cloudinary
 
+## 2.5.33
+
+### Patch Changes
+
+- @dyrected/core@2.5.33
+
 ## 2.5.32
 
 ### Patch Changes